+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Initialising an Element with given meta or attributes as ObjectElement is now
   supported.
+- When converting JavaScript values to Refract, objects are now supported.
 
 # 0.17.0 - 2017-06-16